requirements.txt
numpy
trimesh
Pillow
pandas
scipy